# api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: Odal Node API
  version: 0.12.0
  description: |
    **Sovereign Digital Product Passport Infrastructure**

    The Odal API is split across two deployable services:

    - **Node** (`http://localhost:8001`) — the MVP service that bundles vault,
      identity, and integrator on a single port with sub-path routing:
      `/vault/*`, `/identity/*`, `/integrator/*`.
    - **Resolver** (`http://localhost:8003`) — standalone public resolver for
      QR code scans and supply-chain integrations. Unauthenticated.

    All authenticated vault endpoints are versioned under `/vault/api/v1/`.
    Path parameters use OpenAPI 3.1 `{id}` brace syntax (aligned with
    Axum 0.8 routing).

    Authentication uses `Authorization: Bearer odal_sk_...` API keys (SHA-256
    hashed, prefix-indexed) or HTTP Basic auth for local development.

    Regulation references:
    - EU Ecodesign for Sustainable Products Regulation (ESPR) 2024/1781
    - EU Battery Regulation 2023/1542
    - W3C DID Core Spec 1.0 / W3C VC Data Model 2.0
  contact:
    name: Odal Support
    email: contact@odal-node.io
    url: https://docs.odal-node.io
  license:
    name: BSL-1.1
    url: https://mariadb.com/bsl11/
  x-legal-entity: Odal Node
servers:
  - url: http://localhost:8001
    description: Local development — odal-node (vault + identity public + integrator)
  - url: http://localhost:8002
...
  - url: http://localhost:8003
    description: Local development — odal-resolver (public)
...
paths:
...
  /dpp/{dppId}/qr:
    get:
      operationId: getDppQrCode
      summary: Get QR code for a DPP
      description: |
        Returns a QR code image for the given DPP. The QR encodes the passport's
        GS1 Digital Link on the resolver's configured base
        (e.g. `https://id.odal-node.io/01/{gtin}/21/{serial}`).
        This endpoint is on the **resolver** service (port 8003).
      tags:
        - Public Resolver
      parameters:
        - name: dppId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DppId'
      responses:
        '200':
          description: QR code image.
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /01/{gtin}:
    get:
      operationId: resolveByGtin
      summary: GS1 Digital Link resolver
      description: |
        GS1-conformant resolver (GS1-CRSV1) keyed by GTIN-14. Behaviour depends
        on the `linkType` query / `Accept` header:
        - default → `307` redirect to the HTML DPP page
        - `?linkType=gs1:pip` / `gs1:dpp` (and related) → `307` redirect to the DPP
        - `?linkType=linkset` or `Accept: application/linkset+json` → `200`
          RFC 9264 linkset
      tags:
        - Public Resolver
      parameters:
        - name: gtin
          in: path
          required: true
          schema:
            type: string
            example: '09506000134352'
        - name: linkType
          in: query
          required: false
          schema:
            type: string
          description: e.g. linkset, gs1:pip, gs1:dpp
      responses:
        '200':
          description: RFC 9264 linkset (when a linkset is requested).
          content:
            application/linkset+json:
              schema:
                type: object
        '307':
          description: Redirect to the DPP page (`Location` header).
        '404':
          description: No published DPP for this GTIN, or unknown link type.
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: |
        Machine-to-machine access via API keys. Keys are prefixed with
        `odal_sk_` and sent as `Authorization: Bearer odal_sk_...`.

        Keys are stored as SHA-256 hashes in PostgreSQL with a 12-character
        prefix index for fast lookup. The plain-text key is only returned
        once at creation time via `POST /vault/api/v1/api-keys`.
    BasicAuth:
      type: http
      scheme: basic
      description: |
        The operator's own bootstrap credential, from the `ADMIN_USERNAME` and
        `ADMIN_PASSWORD` environment variables. It mints the first API key on a
        fresh node — before any key exists — and is the lockout-recovery path
        afterwards, since it carries no `keyId` and so can revoke any key.

        Active in every environment where both variables are set; there is no
        production gate. Leave them unset once an API key exists if you do not
        want the path available.

        Reached only via the `Basic` scheme. A `Bearer` token is never matched
        against it, even one carrying the same `base64(user:pass)` payload.
    MutualTLS:
      type: mutualTLS
      description: |
        Client-certificate (mTLS) authentication for service-to-service calls.
        The standalone identity service accepts the internal `/internal/*`
        endpoints only from a client presenting `CN=odal-vault`. The fused
        `dpp-node` never exposes these over the network — it signs in-process.
  schemas:
...
    DppId:
      type: string
      format: uuid
      description: UUID v7 identifier assigned on creation. Embedded in QR codes and public URLs.
      example: 019723f4-1a2b-7c3d-8e4f-5a6b7c8d9e0f
...
    Problem:
      type: object
      description: |
        RFC 7807 / RFC 9457 problem details. The shape
        `dpp-common::http_problem::Problem` produces, served as
        `application/problem+json`.

        `type` is derived from `title`, so each distinct `title` used across the
        codebase is a stable catalogue key that clients may depend on.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Absolute URI identifying the problem type.
          example: https://problems.odal-node.io/not-found
        title:
          type: string
          description: Short human-readable summary of the problem type.
          example: Not Found
        status:
          type: integer
          description: The HTTP status code, mirroring the status line.
          example: 404
        detail:
          type: string
          description: Human-readable explanation for this specific occurrence.
          example: 'No schema for product group ''nosuchsector''. Known product groups: aluminium, battery.'
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this specific occurrence.
...
  responses:
...
    NotFound:
      description: Resource not found within the operator's scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://problems.odal-node.io/not-found
            title: Not Found
            status: 404
            detail: DPP not found.
...